refactor: low-level rotational partition function now computes the ortho- and para- contributions at the same time

+++ plots.py
@@ -4,6 +4,7 @@ import matplotlib
 matplotlib.use("Agg")
 from partition_functions import molecular_hydrogen_energy
 from constants import BOLTZMANN
+from EOS import *
 import numpy as np
 
 
@@ -16,11 +17,24 @@ def do_plots():
 
     kw = {"color": "black"}
     ax[0].plot(Tgrid, e / (BOLTZMANN * Tgrid), **kw)
+
+    eos = EOS(ChemicalState(hydrogen_massfrac=1.0, metallicity=0.0))
+    u = eos.perH_to_permass(0.5 * e)
+    Tu = eos.u_to_temp(u)
+    if not np.isclose(Tgrid, Tu, rtol=1e-15).all():
+        raise ValueError("Inverted temperature calculated from u not consistent.")
+    ax[0].plot(Tu, e / (BOLTZMANN * Tgrid), ls="dotted", color="red")
+
     ax[1].plot(Tgrid, cv / BOLTZMANN, **kw)
     ax[2].plot(Tgrid, gamma, **kw)
 
-    T, gamma = np.loadtxt("data/boley_2007_gamma.dat").T
-    ax[2].plot(T, gamma, color="red", ls="dashed", label="Boley 2007")
+    T, gamma_ref = np.loadtxt("data/boley_2007_gamma.dat").T
+    if not np.isclose(
+        gamma_ref, np.interp(np.log10(T), np.log10(Tgrid), gamma), rtol=1e-2
+    ).all():
+        raise ValueError("Adiabatic index not consistent.")
+
+    ax[2].plot(T, gamma_ref, color="red", ls="dashed", label="Boley 2007")
     ax[2].legend()
     for a in ax:
         a.set(xlabel=r"$T\,\left(\rm K\right)$", xscale="log")
